File Lineage handling for copying update.
File versions are now recoverable through FileLineage.json files. These track who (Mod or Game) owns and has wants to own a file.

The owner is the one who's version is currently deployed ingame. A "want's to own" is either a previous owner who was overridden or a later deployed mod that did not at the time have load priority. They are kept to ensure that removed mods can roll-back a file's version to use previously lower priority mods or overriden mods to correctly own the file.

So far the method to handle copying individual files is only complete. Restoring backups, and rolling back file version and the FileLineageIO methods are not started.

BackupManifests are no longer needed.